Divide classes into packages and rectify command classes
