+++ src/main/java/fairy/common/Messages.java
@@ -1,8 +1,10 @@
+package fairy.common;
+
 public class Messages {
 
     public static final String MESSAGE_ADD_TASK_SUCCESS = "Yes, Master. I've added this task to your list:\n%s\nThere are %d tasks in your list now.";
     public static final String MESSAGE_ARGUMENT_EXCEPTION = "Argument exception: No enough arguments.";
-    public static final String MESSAGE_COMMAND_NOT_FOUND = "Command not found: %s";
+    public static final String MESSAGE_COMMAND_NOT_FOUND = "fairy.command.Command not found: %s";
     public static final String MESSAGE_DATETIME_EXCEPTION = "Date time exception: %s";
     public static final String MESSAGE_DATE_PARSE_EXCEPTION = "Date time exception: Wrong format or illegal time. Correct format: YYYYMMDD";
     public static final String MESSAGE_DATETIME_PARSE_EXCEPTION = "Date time exception: Wrong format or illegal time. Correct format: YYYYMMDD hhmm";
