+++ src/main/java/fairy/common/utils/FairyDateTimeFormatter.java
@@ -1,3 +1,5 @@
+package fairy.common.utils;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
