+++ src/main/java/fairy/exception/EmptyListException.java
@@ -1,3 +1,5 @@
+package fairy.exception;
+
 public class EmptyListException extends Exception{
 
     public EmptyListException(){
