+++ src/main/java/fairy/exception/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package fairy.exception;
+
 public class InvalidCommandException extends Exception{
 
     public InvalidCommandException(){
