+++ src/main/java/fairy/parser/CommandParser.java
@@ -1,3 +1,7 @@
+package fairy.parser;
+
+import fairy.command.Command;
+
 import java.util.ArrayList;
 import java.util.List;
