+++ src/main/java/fairy/task/Deadline.java
@@ -1,3 +1,5 @@
+package fairy.task;
+
 import java.time.LocalDateTime;
 
 public class Deadline extends Task {
