+++ src/main/java/fairy/task/Event.java
@@ -1,3 +1,5 @@
+package fairy.task;
+
 import java.time.LocalDateTime;
 
 public class Event extends Task {
