+++ src/main/java/fairy/task/Task.java
@@ -1,3 +1,5 @@
+package fairy.task;
+
 public class Task {
     private final String taskName;
     private boolean done;
