+++ src/main/java/fairy/task/Todo.java
@@ -1,3 +1,5 @@
+package fairy.task;
+
 public class Todo extends Task {
     public Todo(String taskName) {
         super(taskName);
