+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 
-    Hello, Master. This is Fairy, your personal assistant.
+    Hello, Master. This is fairy.Fairy, your personal assistant.
     How can I help you?
 
 >
